// @types/generated/contentful.d.ts
export interface IAboutPageFields {
  /** Title */
  title?: string | undefined

  /** Hero Image */
  heroImage: Asset

  /** Text Sections */
  textSections: ITextSection[]

  /** Media Links */
  mediaLinks: ILink[]

  /** Staff Section Title */
  staffSectionTitle: string

  /** Staff Section Text */
  staffSectionText: Document

  /** Staff */
  staff: IStaffMember[]
}
// ...
export interface ILinkFields {
  /** Display text */
  displayText: string

  /** href */
  href: string
}

export interface ILink extends Entry<ILinkFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'link'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}
// ...
export interface IStaffMemberFields {
  /** Name */
  name: string

  /** Role */
  role: string

  /** Image */
  image: Asset
}

export interface IStaffMember extends Entry<IStaffMemberFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'staffMember'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}
// ...
export interface ITextSectionFields {
  /** Heading */
  heading: string

  /** Text */
  text: Document
}

/** Reusable text section with a section heading */

export interface ITextSection extends Entry<ITextSectionFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'textSection'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}